+++ Src/z80.h
@@ -107,14 +107,14 @@ extern int8_t gixiyshift;
 extern z80_state_t state;
 extern int z80_tstates;
 
-extern void (*port_out)(register uint16_t addr, register uint8_t data);
-extern uint8_t (*port_in)(register uint16_t addr);
+extern void (*port_out)(register uint32_t addr, register uint32_t data);
+extern uint8_t (*port_in)(register uint32_t addr);
 
-void z80_Init(void (*outfn)(register uint16_t addr, register uint8_t data), uint8_t (*infn)(register uint16_t addr));
+void z80_Init(void (*outfn)(register uint32_t addr, register uint32_t data), uint8_t (*infn)(register uint32_t addr));
 void z80_reset();
 void z80_interrupt();
 void z80_nmi();
-void req_int(register uint8_t type);
+void req_int(register uint32_t type);
 void z80_step();
