+++ Src/zx_system.h
@@ -17,7 +17,7 @@
 extern uint8_t zxsys_isrun;
 
 void zxsys_Run();
-void zxports_out(register uint16_t addr, register uint8_t data);
-uint8_t zxports_in(register uint16_t addr);
+void zxports_out(register uint32_t addr, register uint32_t data);
+uint8_t zxports_in(register uint32_t addr);
 
 #endif /* ZX_SYSTEM_H_ */
